backend/requirements.txt
Django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
dj-database-url==2.2.0
psycopg[binary]==3.2.10
gunicorn==23.0.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
whitenoise==6.8.2
django-storages[boto3]==1.14.4
